🎨 historyHandler util 함수 분리, 선택된 히스토리 표시
