+++ client/src/components/molecules/HistoryWindow/index.tsx
@@ -10,18 +10,24 @@ import { getUser } from 'utils/helpers';
 
 interface IProps {
   onClick: (historyData: IHistoryData, idx: number) => (event: MouseEvent) => void;
+  currentHistory: IHistoryData;
 }
 
-const HistoryWindow: React.FC<IProps> = ({ onClick }) => {
+const HistoryWindow: React.FC<IProps> = ({ onClick, currentHistory }) => {
   const { containerRef, dragRef } = useDragBackground();
   const historyData = useRecoilValue(historyDataState);
   const userList = useRecoilValue(userListState);
 
   return (
     <Wrapper ref={containerRef} className='background'>
-      {historyData && userList
+      {historyData && userList && currentHistory
         ? historyData.map((historyDataPiece, idx) => (
-            <IconWrapper key={idx} onClick={onClick(historyDataPiece, idx)}>
+            <IconWrapper
+              key={idx}
+              onClick={onClick(historyDataPiece, idx)}
+              isSelected={historyDataPiece.historyId === currentHistory.historyId}
+              color={getUser(historyDataPiece.user, userList)!.color}
+            >
               <UserIcon user={getUser(historyDataPiece.user, userList)!} cursor='pointer' />
             </IconWrapper>
           ))
