+++ client/src/components/molecules/HistoryWindow/style.tsx
@@ -1,5 +1,9 @@
 import styled from '@emotion/styled';
 
+interface IProps {
+  isSelected: boolean;
+}
+
 export const Wrapper = styled.div`
   background-color: ${({ theme }) => theme.color.bgWhite};
   width: 100%;
@@ -12,4 +16,12 @@ export const Wrapper = styled.div`
   position: relative;
 `;
 
-export const IconWrapper = styled.div``;
+export const IconWrapper = styled.div<IProps>`
+  ${({ isSelected, theme }) => (isSelected ? { backgroundColor: theme.color.primary1 } : { border: 'none' })};
+  border-radius: 50%;
+  width: 50px;
+  height: 50px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+`;
